Add files via upload
Data Analysis Internship Task 1

Unemployment In India This notebook presents an analysis of unemployment data in India, covering data preprocessing, exploratory data analysis (EDA), hypothesis testing, and model building. The goal is to uncover trends, patterns, and potential predictive insights from the dataset.

Understanding the Problem and Data In this project, we're analyzing unemployment data in India. The dataset contains information such as:

Region: Different regions in India. Date: The time when the unemployment data was recorded. Frequency: The frequency of data recording (e.g., monthly). Estimated Unemployment Rate (%): The percentage of the labor force that is unemployed. Estimated Employed: The estimated number of employed individuals. Estimated Labour Participation Rate (%): The percentage of the population that is either employed or actively looking for work. Area: Whether the area is urban or rural. Goal: Explore this data to identify trends and patterns, and possibly build models to predict future unemployment rates.